Heavy structural refactor
Changes how commands are loaded and executed to, hopefully, make it
easier to build and maintain individual commands.

Created the start of a config management system based heavily on the
config management system provided in Colima. Future functionality aligns
pretty heavily with how Colima operates and the CLI UX there is quite
nice.

Replaced slog with zerolog, which — at the time anyway — was much faster
when logging json and significantly easier to convert to a nice
human-readable format when outputting to console.